+++ package.json
@@ -29,6 +29,7 @@
     "react-scroll": "^1.9.0",
     "react-spring": "^9.7.4",
     "react-toastify": "^10.0.5",
+    "recoil": "^0.7.7",
     "tailwind-scrollbar-hide": "^1.1.7",
     "zod": "^3.23.8"
   },
