+++ src/routes/admin.tsx
@@ -14,58 +14,14 @@ import { getOrganizationConfig } from "../services/common-services";
 import { LoadingOverlay } from "@mantine/core";
 import Loader from "../components/common/loader/loader";
 import AdminProfile from "../components/admin/dashboard/profile/profile";
+import { useSetRecoilState, useRecoilValue } from "recoil";
+import { organizationThemeAtom } from "../atoms/organization-atom";
 
 const AdminRoutes = () => {
   const { organization } = useParams<{ organization: string }>();
   const [isLoading, setIsLoading] = useState<boolean>(false);
-  const [organizationConfig, setOrganizationConfig] =
-    useState<OrganizationConfig>({
-      organization_name: "default",
-      organization_theme: {
-        logo: "/data-store.png",
-        organization: "default",
-        theme: {
-          primaryColor: "primary",
-          colorScheme: "dark",
-          fontFamily: "Arial, sans-serif",
-          button: {
-            color: "#343a40", // Dark gray
-            textColor: "#ffffff", // White for contrast
-          },
-          colors: {
-            primary: [
-              "#343a40", // Dark gray
-              "#2c3136", // Slightly darker gray
-              "#23272b", // Darker gray
-              "#1d2124", // Even darker gray
-              "#16191c", // Almost black gray
-              "#0f1214", // Near black
-              "#080a0b", // Very dark
-              "#030405", // Almost completely black
-              "#000000", // Black
-              "#000000", // Black
-            ],
-            secondary: [
-              "#adb5bd", // Muted gray
-              "#949aa0", // Slightly darker gray
-              "#7b8287", // Darker gray
-              "#62696f", // Even darker gray
-              "#4a5157", // Very dark gray
-              "#32383e", // Almost black
-              "#1a1f24", // Near black
-              "#080a0b", // Very dark
-              "#000000", // Black
-              "#000000", // Black
-            ],
-          },
-          color: "#ffffff", // White text
-          backgroundColor: "#1b1e21", // Dark background
-          borderColor: "#4a4e69", // Muted purple for borders
-          linkColor: "#ff4d77", // Bright pink for links
-          headerBackgroundColor: "#23272b", // Dark header background
-        },
-      },
-    });
+  const organizationConfig = useRecoilValue(organizationThemeAtom);
+  const setOrganizationConfig = useSetRecoilState(organizationThemeAtom);
 
   useEffect(() => {
     if (organization) {
@@ -77,7 +33,7 @@ const AdminRoutes = () => {
         }
       );
     }
-  }, []);
+  }, [organization, setOrganizationConfig]);
 
   const theme = {
     colorScheme: organizationConfig.organization_theme.theme.colorScheme,
@@ -142,19 +98,9 @@ const AdminRoutes = () => {
         loaderProps={{ children: <Loader /> }}
       />
       <Routes>
-        <Route
-          path="/login"
-          element={<AdminLogin organizationConfig={organizationConfig} />}
-        />
-        <Route
-          element={
-            <AdminProtectedRoutes organizationConfig={organizationConfig} />
-          }
-        >
-          <Route
-            path="/dashboard"
-            element={<AdminDashboard organizationConfig={organizationConfig} />}
-          >
+        <Route path="/login" element={<AdminLogin />} />
+        <Route element={<AdminProtectedRoutes />}>
+          <Route path="/dashboard" element={<AdminDashboard />}>
             <Route
               path="addemployee"
               element={<AddEmployee organizationConfig={organizationConfig} />}
@@ -180,13 +126,10 @@ const AdminRoutes = () => {
   );
 };
 
-const AdminProtectedRoutes = ({
-  organizationConfig,
-}: {
-  organizationConfig: OrganizationConfig;
-}) => {
+const AdminProtectedRoutes = () => {
   const token = localStorage.getItem("adminToken");
   const userRole = localStorage.getItem("userRole");
+  const organizationConfig = useRecoilValue(organizationThemeAtom);
   const navigate = useNavigate();
 
   useEffect(() => {
